Small tweaks to README.md
Thank you so much for the reference to Titaniumifier! Very appreciated!

I tweaked it a little to make it clear that Titaniumifier must be used in Host-mode to enable Babel transpilation.

Also I added a small banner for the [discussion](https://github.com/TiForward/discuss/issues/8) that’s going on Ti.Forward on the support for ES2015.

Again, thank you!